api/requirements.txt
flask
mariadb
python-dotenv
simplejson
gunicorn